Added initial multiple-home support.
We are using the standard CommandBook homes.csv format still, to keep backwards compatibility. However, we are using an invalid Minecraft username (containing @) to store the additional homes.
The new permission commandbook.home.set.multiple has been added to allow for permission control of homes.

There are still a few features left to be desired, one being that there is no way to list a user's home(s).